created new system for upload without admin-panel

+++ src/videoflix_db/signals.py
@@ -15,8 +15,8 @@ def video_post_save(sender, instance, created, **kwargs):
 
 @receiver(post_delete, sender=Video)
 def auto_delete_video_on_delete(sender, instance, **kwargs):
-    fields = ['file1080p', 'file720p', 'file360p', 'file240p', 'preview144p', '_preview', '_big']
+    fields = ['file1080p', 'file720p', 'file360p', 'file240p', 'filePreview144p', 'image', 'bigImage']
     for field in fields:
         file_field = getattr(instance, field)
-        if file_field and os.path.isfile(file_field.path):
+        if file_field and hasattr(file_field, 'path') and os.path.isfile(file_field.path):
             os.remove(file_field.path)
